refactor: 토큰 발급 날짜를 인자로 받게 변경
